Title: DE1050 - Hadoop core package installs sometimes fails when Cloudera Manager is assigned to the edgenode.
Symptoms:
The CM Host manager reports bad user and groups for ozzie, hue and hive. All Hadoop core
and eco-systems packages are not installed. The problem does not always happen and it is
timing/sequence related.

Notes:
The cm-agent process was starting up on the same node as the cm-server and the cm-agent
was pre-configured (default package install) to use "localhost"  as the target cm-server
host name (/etc/Cloudera-scm-agent/config.ini).  CM has logic in it to skip the package
installation process (as a whole) if it sees the cm-agent heartbeats across the network
for that particular node. This resulted in the package install process being skipped
(all packages) for that node depending on if the agent was up and running at the point
in time when the CM package install process was invoked. Solution: Ensure that the
cm-agent process is not running with a valid cm-server host until CM actually deploys
the packages and activates the cm-agent configuration.